concierge/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
grpcio-tools~=1.30
